+++ src/ProjectService.Mappers/Models/Interfaces/IDepartmentInfoMapper.cs
@@ -2,7 +2,7 @@
 using LT.DigitalOffice.Kernel.Attributes;
 using LT.DigitalOffice.ProjectService.Models.Dto.Models;
 
-namespace LT.DigitalOffice.ProjectService.Mappers.ModelsMappers.Interfaces
+namespace LT.DigitalOffice.ProjectService.Mappers.Models.Interfaces
 {
     [AutoInject]
     public interface IDepartmentInfoMapper
