+++ src/ProjectService.Mappers/Models/Interfaces/IProjectUserInfoMapper.cs
@@ -3,7 +3,7 @@ using LT.DigitalOffice.Kernel.Attributes;
 using LT.DigitalOffice.ProjectService.Models.Db;
 using LT.DigitalOffice.ProjectService.Models.Dto.Models.ProjectUser;
 
-namespace LT.DigitalOffice.ProjectService.Mappers.ModelsMappers.Interfaces
+namespace LT.DigitalOffice.ProjectService.Mappers.Models.Interfaces
 {
     [AutoInject]
     public interface IProjectUserInfoMapper
