+++ src/ProjectService.Mappers/Models/ProjectUserInfoMapper.cs
@@ -1,5 +1,5 @@
 ﻿using LT.DigitalOffice.Broker.Models;
-using LT.DigitalOffice.ProjectService.Mappers.ModelsMappers.Interfaces;
+using LT.DigitalOffice.ProjectService.Mappers.Models.Interfaces;
 using LT.DigitalOffice.ProjectService.Models.Db;
 using LT.DigitalOffice.ProjectService.Models.Dto.Enums;
 using LT.DigitalOffice.ProjectService.Models.Dto.Models.ProjectUser;
