+++ src/ProjectService.Mappers/Models/TaskInfoMapper.cs
@@ -1,10 +1,10 @@
 ﻿using LT.DigitalOffice.Broker.Models;
-using LT.DigitalOffice.ProjectService.Mappers.ModelsMappers.Interfaces;
+using LT.DigitalOffice.ProjectService.Mappers.Models.Interfaces;
 using LT.DigitalOffice.ProjectService.Models.Db;
 using LT.DigitalOffice.ProjectService.Models.Dto.Models;
 using System;
 
-namespace LT.DigitalOffice.ProjectService.Mappers.ModelsMappers
+namespace LT.DigitalOffice.ProjectService.Mappers.Models
 {
     public class TaskInfoMapper : ITaskInfoMapper
     {
