+++ src/ProjectService.Mappers/Responses/FindProjectsResponseMapper.cs
@@ -1,12 +1,12 @@
-﻿using LT.DigitalOffice.ProjectService.Mappers.ModelsMappers.Interfaces;
-using LT.DigitalOffice.ProjectService.Mappers.ResponsesMappers.Interfaces;
+﻿using LT.DigitalOffice.ProjectService.Mappers.Models.Interfaces;
+using LT.DigitalOffice.ProjectService.Mappers.Responses.Interfaces;
 using LT.DigitalOffice.ProjectService.Models.Db;
 using LT.DigitalOffice.ProjectService.Models.Dto.Models;
 using LT.DigitalOffice.ProjectService.Models.Dto.ResponsesModels;
 using System;
 using System.Collections.Generic;
 
-namespace LT.DigitalOffice.ProjectService.Mappers.ResponsesMappers
+namespace LT.DigitalOffice.ProjectService.Mappers.Responses
 {
     public class FindProjectsResponseMapper : IFindProjectsResponseMapper
     {
