+++ src/ProjectService.Mappers/Responses/ProjectExpandedResponseMapper.cs
@@ -1,5 +1,5 @@
-﻿using LT.DigitalOffice.ProjectService.Mappers.ModelsMappers.Interfaces;
-using LT.DigitalOffice.ProjectService.Mappers.ResponsesMappers.Interfaces;
+﻿using LT.DigitalOffice.ProjectService.Mappers.Models.Interfaces;
+using LT.DigitalOffice.ProjectService.Mappers.Responses.Interfaces;
 using LT.DigitalOffice.ProjectService.Models.Db;
 using LT.DigitalOffice.ProjectService.Models.Dto.Models;
 using LT.DigitalOffice.ProjectService.Models.Dto.Models.ProjectUser;
@@ -7,7 +7,7 @@ using LT.DigitalOffice.ProjectService.Models.Dto.Responses;
 using System;
 using System.Collections.Generic;
 
-namespace LT.DigitalOffice.ProjectService.Mappers.ResponsesMappers
+namespace LT.DigitalOffice.ProjectService.Mappers.Responses
 {
     public class ProjectExpandedResponseMapper : IProjectExpandedResponseMapper
     {
