+++ test/ProjectService.Mappers.UnitTests/Models/ProjectFileInfoMapperTests.cs
@@ -1,5 +1,5 @@
-﻿using LT.DigitalOffice.ProjectService.Mappers.ModelsMappers;
-using LT.DigitalOffice.ProjectService.Mappers.ModelsMappers.Interfaces;
+﻿using LT.DigitalOffice.ProjectService.Mappers.Models;
+using LT.DigitalOffice.ProjectService.Mappers.Models.Interfaces;
 using LT.DigitalOffice.ProjectService.Models.Db;
 using LT.DigitalOffice.ProjectService.Models.Dto.Models;
 using LT.DigitalOffice.UnitTestKernel;
