+++ test/ProjectService.Mappers.UnitTests/Models/ProjectInfoMapperTests.cs
@@ -1,6 +1,6 @@
 ﻿using LT.DigitalOffice.Broker.Responses;
-using LT.DigitalOffice.ProjectService.Mappers.ModelsMappers;
-using LT.DigitalOffice.ProjectService.Mappers.ModelsMappers.Interfaces;
+using LT.DigitalOffice.ProjectService.Mappers.Models;
+using LT.DigitalOffice.ProjectService.Mappers.Models.Interfaces;
 using LT.DigitalOffice.ProjectService.Models.Db;
 using LT.DigitalOffice.ProjectService.Models.Dto.Enums;
 using LT.DigitalOffice.ProjectService.Models.Dto.Models;
