+++ test/ProjectService.Mappers.UnitTests/Models/ProjectUserInfoMapperTests.cs
@@ -1,6 +1,6 @@
 ﻿using LT.DigitalOffice.Broker.Models;
 using LT.DigitalOffice.ProjectService.Mappers.Interfaces;
-using LT.DigitalOffice.ProjectService.Mappers.ModelsMappers.Interfaces;
+using LT.DigitalOffice.ProjectService.Mappers.Models.Interfaces;
 using LT.DigitalOffice.ProjectService.Models.Db;
 using LT.DigitalOffice.ProjectService.Models.Dto.Models.ProjectUser;
 using LT.DigitalOffice.UnitTestKernel;
