+++ test/ProjectService.Mappers.UnitTests/Responses/FindProjectsResponseMapperTests.cs
@@ -1,6 +1,6 @@
-﻿using LT.DigitalOffice.ProjectService.Mappers.ModelsMappers.Interfaces;
-using LT.DigitalOffice.ProjectService.Mappers.ResponsesMappers;
-using LT.DigitalOffice.ProjectService.Mappers.ResponsesMappers.Interfaces;
+﻿using LT.DigitalOffice.ProjectService.Mappers.Models.Interfaces;
+using LT.DigitalOffice.ProjectService.Mappers.Responses;
+using LT.DigitalOffice.ProjectService.Mappers.Responses.Interfaces;
 using LT.DigitalOffice.ProjectService.Models.Db;
 using LT.DigitalOffice.ProjectService.Models.Dto.Models;
 using LT.DigitalOffice.ProjectService.Models.Dto.ResponsesModels;
@@ -10,7 +10,7 @@ using NUnit.Framework;
 using System;
 using System.Collections.Generic;
 
-namespace LT.DigitalOffice.ProjectService.Mappers.UnitTests.ResponseMappers
+namespace LT.DigitalOffice.ProjectService.Mappers.UnitTests.Responses
 {
     class FindProjectsResponseMapperTests
     {
